Refactored & Fixed circular motion Issue @ Backend
