--- src/redux/features/BookMark/bookMarkSlice.js
@@ -1,18 +1,0 @@
-import { createSlice } from "@reduxjs/toolkit";
-import { addToBookmarks } from "./bookmarkThunk";
-
-const initialState = {
-  bookmarks: [],
-};
-
-const bookMarkSlice = createSlice({
-  name: "bookmark",
-  initialState,
-  extraReducers: (builder) => {
-    builder.addCase(addToBookmarks.fulfilled, (state, action) => {
-      state.bookmarks = action.payload.bookmarks;
-    });
-  },
-});
-
-export default bookMarkSlice.reducer;
