--- src/redux/features/allPostsSlice.js
@@ -1,16 +1,0 @@
-import { createSlice } from "@reduxjs/toolkit";
-
-const initialState = { allPosts: [] };
-
-const allPostsSlice = createSlice({
-  name: "allPosts",
-  initialState,
-  reducers: {
-    setAllPosts: (state, action) => {
-      state.allPosts = action.payload;
-    },
-  },
-});
-
-export const { setAllPosts } = allPostsSlice.actions;
-export default allPostsSlice.reducer;
