+++ src/redux/features/post/postThunk.js
@@ -92,3 +92,21 @@ export const dislikePost = createAsyncThunk(
     }
   }
 );
+
+export const addToBookmarks = createAsyncThunk(
+  "post/addToBookmarks",
+  async (postId, { rejectWithValue, getState }) => {
+    try {
+      const response = await axios.post(
+        `/api/users/bookmark/${postId}`,
+        {},
+        {
+          headers: { authorization: getState().auth.token },
+        }
+      );
+      return response.data;
+    } catch (error) {
+      return rejectWithValue(error.response.data);
+    }
+  }
+);
